src/backend-django/requirements.txt
Django==5.0.8
gunicorn==22.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
django-ratelimit==4.1.0
django-extensions==3.2.3
numpy>=1.26.0,<2.0
django-filter==24.2
djangorestframework-simplejwt==5.3.0
Pillow>=9.5.0

# Updated 2025-01-10 - CSS Design System Integration
# Backend optimizations for enhanced frontend experience